requirements.txt
opencv-python
requests
python-dotenv
Pillow
SpeechRecognition
pyttsx3